requirements.txt
requests
progress
python-telegram-bot
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
telegraph
psycopg2-binary